+++ src/tests/Pipelinez.SqlServer.Tests/EndToEnd/Models/TestSqlServerRecord.cs
@@ -1,0 +1,10 @@
+using Pipelinez.Core.Record;
+
+namespace Pipelinez.SqlServer.Tests.EndToEnd.Models;
+
+public sealed class TestSqlServerRecord : PipelineRecord
+{
+    public required string Id { get; init; }
+
+    public required string Value { get; set; }
+}